requirements.txt
requests
datetime
python-dateutil
urllib3
configparser